backend/requirements.txt
alembic==1.7.5
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
crypto==1.4.1
cryptography==36.0.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
mysqlclient==2.1.0
Naked==0.1.31
protobuf==3.19.3
psycopg2-binary==2.9.3
pycparser==2.21
pycrypto==2.6.1
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.29
twilio==7.5.0
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
